popravljena responzivnost i donekle sredjeni licni podaci
